requirements.txt
drfdocs
django-simple-email-confirmation